added alternative naming, new function for finding added elib id, added colouring functinality for people with similar names

+++ parsers_two_table/findPeopleWithoutID.py
@@ -1,10 +1,15 @@
 import pandas as pd
 from sqlalchemy import create_engine
 from dbsettings import database_parametres
+from find_similar_fullnames import find_similar_fullnames
+from apply_colours_to_excel import apply_fill_colors
+from find_alternative_names import merge_authors_by_enter_id
 import random
 import os
 import time
 import psutil
+
+
 
 def generate_unique_id(author_id_list):
     while True:
@@ -43,10 +48,9 @@ def update_author_id(excel_file_path):
                     return ' '
 
             df_null['possible_id_from_xml'] = df_null.apply(compute_possible_id, axis=1)
+
         if len(filtered_rows_from_db) > 0:
                 filtered_rows_from_db = filtered_rows_from_db.drop_duplicates(subset=['author_fullname'], keep='last')
-
-
                 author_name_id_dict = dict(zip(filtered_rows_from_db['author_fullname'], filtered_rows_from_db['author_id']))
 
                 def compute_possible_id(row):
@@ -56,13 +60,24 @@ def update_author_id(excel_file_path):
                         return ' '
 
                 df_null['possible_id_from_db'] = df_null.apply(compute_possible_id, axis=1)
+        df_null['enter_id'] = ''
+        df_null.to_excel('author_filtered_data.xlsx')
     except Exception as e:
         print("An error occurred:", str(e))
-    df_null['enter_id'] = ''
-    df_null.to_excel('author_filtered_data.xlsx')
     def check_author_id(file_path):
         df = pd.read_excel(file_path)
         return all(df['enter_id'] != ' ')
+
+
+    if 'possible_id_from_db' in df_null.columns and os.path.exists('alternative_names.xlsx'):
+        print('exists')
+        df_null = pd.read_excel('author_filtered_data.xlsx', index_col=0)
+        df_alternative = pd.read_excel('alternative_names.xlsx')
+        df_null['alternative_name'] = df_null['possible_id_from_db'].map(
+        df_alternative.set_index('enter_id')['new_column_name'])
+        df_null.to_excel('author_filtered_data.xlsx')
+    find_similar_fullnames('author_filtered_data.xlsx')
+    apply_fill_colors('author_filtered_data.xlsx')
     while True:
         os.system(f'start excel author_filtered_data.xlsx')
         while True:
@@ -91,7 +106,8 @@ def update_author_id(excel_file_path):
     df_null2 = pd.read_excel('author_filtered_data.xlsx', index_col=0)
     df['author_id'] = df.apply(update_author_id, axis=1)
     df.to_excel('authors_organisations.xlsx')
+    merge_authors_by_enter_id('author_filtered_data.xlsx', 'alternative_names.xlsx')
 
 if __name__ == "__main__":
-    update_author_id('authors_organisations.xlsx')
+    update_author_id('../authors_organisations.xlsx')
